[QEMU] Fix QEMU boot issue in nested virtualization env
Is is reported that SBL QEMU target crashes when running in nested
virtualization env. For example, run QEMU with KVM flag inside
Ubuntu running on Windows Virtual Box.
After debug, it turned out somehow the KVM generated exception for
instruction "movd mm0, eax" at the very beginning. This was used to
save the CPU BIST info. If it is moved to be after switching to FLAT32
mode, then everything goes well.  This patch moved the BIST saving
into later stage in the boot in order to address the KVM crash issue.

This fixed #1469.

Signed-off-by: Maurice Ma <maurice.ma@intel.com>