Add conditional scope for I2C pad and panel
Current SBL ACPI table does not define any I2C pad or panel. And
it will cause ACPI error for Linux. This patch added conditional
scope for I2C pad and panel reference so that if pad type or panel
type is not defined, these scope will not be used by ACPI.

This fixed #1496.

Signed-off-by: Maurice Ma <maurice.ma@intel.com>