Add capsule update from multiple USB devices
This patch will try to find capsule image from each of connected USB
devices if the HwPart of CapsuleInformation is set to 0xFF.

When multiple USB devices are attached, current SBL will try to
find the capsule image from the device with index specified by HwPart
in the CapsuleInformation option. However, it is hard to determine
the USB device index order since it depends on which port the device
is connected to.

This patch extends the feature of "boot from multiple USB device" from
OS Loader to capsule update.

Verified: TGL-UP3 RVP

Signed-off-by: Stanley Chang <stanley.chang@intel.com>