fix the wrong sblopen_dir value in StitchLoader.py and StitchIfwi.py
This patch can reduce the chance of setting SBL_SOURCE when the
SBL root directory is not named as "SblOpen" or when the scripts
are run from the path other than the SBL root directory. Also note
that os.path.abspath() returns the absolute path relative to the
current working directory instead of the real path of __file__.
So in StitchIfwi.py, sblopen_dir was incorrect since the working
directory had been changed before calling stitch().

Signed-off-by: Vincent Chen <vincent.chen@intel.com>