Revert "[UPX] Disable malfunctioned USB2 port"
This reverts commit 1caacefeb508a41c4363e64ce45159b72f4b2f73.

The USB port that does not respond to USB bus enumeration is port 8 on
the newer UPX board. The old UPX board might have different behavior.
The original commit was valid on old UPX board only, and this patch
reverted it. Confirmed the USB error message disappeared on the new
UPX board after the reverting.

Signed-off-by: Maurice Ma <mauricemx.ma@gmail.com>