Add GpioGetStrapValue() helper function
This patch adds a new GpioGetStrapValue() helper function which allows
reading the pin-strap value of GPIO pins.  This is useful for reading
the value of hard straps or of BOARD_ID pins or of other GPIOs for
which the input buffer may be disabled, as the pin-strap value bit is
valid even when a pin's input buffer has been disabled.

Signed-off-by: Lennert Buytenhek <buytenh@arista.com>