Allow multiple SBL components to be updated from the same FWU capsule
Currently, after a single redundant SBL component gets updated,
SBL expects there to be no more redundant components left for update.
This made sense as the only redundant component allowed to be
updated used to be the entire BIOS region. Now that the ACM and uCode
redundant regions are allowed to be updated, we need to reset
the state machine after a full redundant component update flow occurs
so that the next one can be triggered.

Signed-off-by: Sean McGinn <sean.mcginn@intel.com>