Check that uCode slot sizes line up during uCode update
Currently, users are able to update the uCode component
with a new binary that uses a different slot size than
the existing binary. If these slot sizes do not line up,
there is potential to trigger recovery (if enabled) or
brick the system. This change disallows updates where
these slot sizes do not line up.

Signed-off-by: Sean McGinn <sean.mcginn@intel.com>