Start TCO timer in SG1A instead of SG1B when resiliency enabled
Previously, the TCO timer was started in SG1B whenever
resiliency was enabled. This was done as it was initially
thought that the TCO timer is not available until SG1B.
However, after some experimentation, it was realized that
the TCO timer is available in SG1A. This change moves the
starting of the TCO timer to SG1A so that it too can also
recover from dead loops, crashes, etc.

Signed-off-by: Sean McGinn <sean.mcginn@intel.com>