Force redundant component check on every FW update iteration
Currently, when a power failure occurs during update of a
BP1 component, the corresponding BP0 component does not get
updated on next boot. This is because the redundant component
check only occurs at the very beginning of the FW update process.

This change corrects this behavior, so that the BP1 component gets
updated along with its corresponding BP0 component on boot after
BP1 component update with power failure.

Signed-off-by: Sean McGinn <sean.mcginn@intel.com>