Move TCO timer start and TCO status check to SG1B
It was discovered that the ACM active timer monitors
boot progress across all of SG1A and the beginning of
SG1B (through FSP-M), flipping the TS if it times out.

This change makes it so that the 2 timers have less
overlap.

This change also consolidates the set and start TCO timer
APIs as well as renames the TCO status check API.

Signed-off-by: Sean McGinn <sean.mcginn@intel.com>