Cover FW_UPDATE_SM_PART_AB state in FwuTopSwapSetting for ADL
FW_UPDATE_SM_PART_AB state occurs when we have successfully updated
both BP0 and BP1. It signifies that that we need to flip the TS
and set FW_UPDATE_SM_DONE state.

It is possible for SBL to get stuck in a FW_UPDATE_SM_PART_AB state if
power failure occurs when its in this state.

This change prevents this from happening.

Signed-off-by: Sean McGinn <sean.mcginn@intel.com>