Remove unnecessary SM Bus in-use status clear
Previously, the reload of the TCO timer blocked the
initialization of the platform ID via the SM Bus in-use
status. Thus, we explicitly cleared it. However, now
that the reload of the TCO timer occurs after the
initialization of the platform ID, this clear is no
longer necessary.

This change removes it.

Signed-off-by: Sean McGinn <sean.mcginn@intel.com>