Fix bound range exception thrown in HECI call for FW capabilities
Currently, whenever a HOB is not available containing FW
capabilities, a HECI call is made for the same. However,
whenever this HECI call is made, it throws a bound range
exception. This is because the ACK message struct used
has a data member but not a MKHI header member,
and the HECI call expects a struct with both members.

Signed-off-by: Sean McGinn <sean.mcginn@intel.com>