Get TPM2 event log working with SBL
For TPM2 event logging, whenever a EV_EFI_VARIABLE_DRIVER_CONFIG
type event is logged, the associated data is expected to be
structured as UEFI_VARIABLE_DATA. However, for key hash and
configuration data, the event data provided was given as a string.

This change corrects the event type for these regions as they
are not actually EFI variables.

This change also provides event data in the structure associated with
the correct event type.

Signed-off-by: Sean McGinn <sean.mcginn@intel.com>