feat: enhance ConfigEditor to support board extension yaml files
When ConfigEditor opens a raw CfgDataDef.yaml, the GPIO GrpIdx
remains a default mapping, which may be undesired. Only after
a DLT file is loaded, it will know which board is used based
on PlatformName. This patch enhances ConfigEditor to reload
CfgDataDef.yaml based on the board name.

This patch also fixes the issue:
- When ConfigEditor opens CfgDataDef.yaml more than once, the
  config changes will not be caught by saving to a DLT file.
  (Save Config Changes to Delta File)
It is because the "info" argument of build_cfg_list() in
GenCfgData.py will retain its value in the lifetime of
ConfigEditor, and the offset will overflow. So need to reset
the offset for a fresh load_yaml().

Signed-off-by: Vincent Chen <vincent.chen@intel.com>