fix: [TGL] FSP returns incorrect number of phases for SiInit (#2181)
MultiPhaseSiInit was fixed in pull #2171 to correctly interpret
NumberOfPhases as 1-based instead of zero based. This exposed an issue
in TGL where FSP returns that there is 1 phase supported in MultiPhaseSi
init, but it actually supports zero phases. The bug fixed by #2171 was
masking this issue.

The workaround in TGL is to override the FSP-S Info header so it appears
to SBL that MultiPhaseSiInit isn't supported (since it isn't supported
properly) so the MultiPhaseSiInit is skipped entirely.

Signed-off-by: Bejean Mosher <bejean.mosher@intel.com>