fix: improve the DX for RsvdSmbusAddressTablePtr #2207
The patch improves the dev experience for configuring FSP-M UPDs:
RsvdSmbusAddressTablePtr and PchNumRsvdSmbusAddresses: Prior to this patch,
developers had to apply a memory patch at address 0xfffc3ff8, which was
cumbersome and inconvenient. With the introduction of this patch, developers
can now directly modify the DLT, simplifying the configuration steps and
improving workflow efficiency.

Signed-off-by: Stanley Chang <stanley.chang@intel.com>