fix: [ASL] remove deprecated PLATFORM_ASL
The PLATFORM_{NAME} macro is dynamically defined by BuildUtility.py, which
is based on the board file name. Since ASL has been merged into ADL-N, the
PLATFORM_ASL macro has become obsolete and is no longer valid. This commit
removes all instances of the deprecated PLATFORM_ASL.

Additionally, this commit addresses an issue where PLATFORM_ADLN was not
defined in Stage1ABoardInitLib.c during the build process for ADL-N. The cause
was that ConfigDataStruct.h was not included.

Signed-off-by: Stanley Chang <stanley.chang@intel.com>