fix: [Common] Assign PCI_MAX_BUS to BusLimit for BusScanTypeList type (#2315)
When platform chooses BusScanTypeList for PCI enumeration, the BusLimit should be set to PCI_MAX_BUS.
The BusScanItems defines primary bus for PCie root port. If there is device attached to root port,
which will present on secondary bus, in the meantime the bus number is greater than Buslimit, this device
cannot be found from enumeration. Assign PCI_MAX_BUS to BusLimit to prevent this issue.

Signed-off-by: Kevin Tsai <kevin.tsai@intel.com>