QEMU FSP BuildTools fix in newer toolchains
When building using newer toolchains, there are a few changes
required that were already patched in EDK2.

Those changes are documented in the two links from EDK2
group discussion:
- https://edk2.groups.io/g/devel/topic/89997410
- https://edk2.groups.io/g/devel/topic/89997412

This set of small changes will later be helpful when
moving to newer OSes and toolchains.
There is no downside in applying them now as the build
in current supported OSes is working fine. It should
be a no harm change with future benefit.

Signed-off-by: Fernando Silva <eng.fernandosilva@outlook.com>