[TGL][TXT] Improved cmos logic for TXT disable usecase
When TXT Feature is disabled TXT_CMOS_STATUS_REG should
be set to disable so that ACM does not try to enable TXT.

This was not happening when previously TXT was enabled
and then new ifwi was flashed with TXT feature disabled.

This is fixed by modifying the if condition in which the
CMOS is cleared.

Signed-off-by: Antara Borwankar <antara.borwankar@intel.com>