feat: [Common] Add support to load payload from PDR
Added support to load EPayload from Platform Data Region (PDR) as an
alternate loading method. This is useful if the overall bootloader
binary cannot fit inside the BIOS region or there is a need to load
multiple payloads or load an OS as a payload. To enable this feature,
user should set self.ENABLE_PDR_LOAD to 1 in the corresponding BoardConfig
file and place the payload at offset 0 in the PDR.

Signed-off-by: Sachin Kamat <sachin.kamat@intel.com>