fix: [ADL] ACPI Gpio Lib
This commit addresses the inconsistent index used by GNUM for ADL-S and
ADL-P/ADL-N. The GPIO_DRIVER_PIN_BASE_NUMBER serves as the index.
For ADL-P/ADL-N, it is located in Field8 of the GPCL package. However, prior
to this commit, ADL-S used Field6 in GPCS.

This update introduces Field6 (PADLOCK offset) and Field7 (PADLOCKTX offset)
for ADL-S and sets GPIO_DRIVER_PIN_BASE_NUMBER to 8. Additionally, the commit
incorporates two ACPI methods, GLOC and GLOT, to read the newly added fields.

Signed-off-by: Stanley Chang <stanley.chang@intel.com>