feat: Add debug messages for PCIe BAR Size calculation
This update addresses issues related to high/low MMIO, which are common in x86
builds, particularly for legacy platforms where x86 is the default build for SBL.

The commit introduces a straightforward debug log to assist in quickly identifying
devices that require larger MMIO space in the low MMIO region.

To enable this feature, please set DEBUG_PCI_ENUM to 1 in PciEnumerationLib.c.

The output log would like:

  PciBus: Discovered PCI @ [00|0A|00]
      VID|DID: 8086|A77D
      Bar: 0 BarType: 3 Size: 0x8000

Signed-off-by: Stanley Chang <stanley.chang@intel.com>