workaround: [UART] ACPI _UID Compatibility
The ACPI spec allows the _UID return value to be either numeric or a string
when the HID is not ACPI0007. However, some legacy UART drivers may only
evaluate it as a numeric value, potentially leading to misinterpretation of
the _UID. To ensure compatibility with these legacy UART drivers, this commit
modifies the _UID of all UART devices from a string to a numeric value.

Additionally, to prevent conflicts with other predefined ACPI devices using _HID
"PNP0501" or "PNP0C02" in SBL, such as H8S2113Sio, the _UID for UART devices
is set to start from 0x50 or 0x60.

Reference: https://uefi.org/specs/ACPI/6.6/06_Device_Configuration.html#uid-unique-id

fixes #2566

Signed-off-by: Stanley Chang <stanley.chang@intel.com>