feat: Enable FSP-S PE/TE Rebase based on board configuration
FSP Spec says FSP-S must be rebased when executed from address other than
its preferred address. However, SBL was not doing this since it is not
required in most cases. Recent changes to FSP-S SEC in latest EDK2
IntelFsp2Pkg cause this to be necessary when FSP is built with GCC in
some cases.

Signed-off-by: Bejean Mosher <bejean.mosher@intel.com>